requirements.txt
requests
future
virtualenv
python-restcountries